spec: a round loop is read wherever it lives, and the refold move is verbatim
